Add multilingual engine bugfix
For multilingual models, we adjust the vocab/embed_matrix size
to make all models/checkpoints fit into the same instance of
`TransformerModel`. This is currently implemented by modifying
the embedding weight matrix to be the size of embedding matrix
of the model with the largest vocabulary. The 'added' words all
of which are semi-treated as paddings, have their word vectors
zerod.

The case that currently is not handled, is the output. The logits
for all of the added words will be 0, but ARE NOT explicitliy treated
as infeasible to predict (in the same way that you would treat <PAD>),
and as such if all other logits are assigned a negative value, the
words in the extended vobulary with logit values 0 now become the most
likely. This obviously causes gibberish translations.

This is non-trivial to fix elegantly, as we're already in a slightly
abusive terrority, in terms of usage of `fairseq`, but the following
changes to `decoder.py` takes a stab at it, by overriding the
`TransformerModel`'s `get_normalized_probs` function, to mask out all
the added vocabulary words, by setting their log probability to
-math.inf.